# openapi.yaml
openapi: 3.0.0
info:
  title: Bank API
  version: 1.0.0
  description: Bank API with interoperability features
  contact:
    name: API Support
    email: support@eerovallistu.site
servers:
  - url: /
    description: Current server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Account:
      type: object
      properties:
        id:
          type: integer
          description: Account ID
        accountNumber:
          type: string
          description: Unique account number with bank prefix
        userId:
          type: integer
          description: ID of the account owner
        balance:
          type: number
          format: float
          description: Current account balance
        currency:
          type: string
          enum: [EUR, USD, GBP]
          description: Account currency
        name:
          type: string
          description: Account name/description
        createdAt:
          type: string
          format: date-time
          description: Account creation date
      required:
        - accountNumber
        - userId
        - currency
        - name
...
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                example: error
              message:
                type: string
                example: Authentication required
    ValidationError:
      description: Input validation error
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                example: error
              errors:
                type: array
                items:
                  type: object
                  properties:
                    field:
                      type: string
                    message:
                      type: string
security:
  - bearerAuth: []
paths:
...
  /accounts:
    get:
      summary: Get all accounts for current user
      tags: [Accounts]
      parameters:
        - in: query
          name: currency
          schema:
            type: string
          description: Filter by currency (EUR, USD, GBP)
        - in: query
          name: sort
          schema:
            type: string
            enum: [balance, -balance, name, -name, createdAt, -createdAt]
          description: Sort accounts (prefix with - for descending)
      responses:
        '200':
          description: List of user accounts
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Account'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    
    post:
      summary: Create a new account
      tags: [Accounts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - currency
                - name
              properties:
                currency:
                  type: string
                  enum: [EUR, USD, GBP]
                  example: EUR
                name:
                  type: string
                  example: Main Savings
      responses:
        '201':
          description: Account created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    $ref: '#/components/schemas/Account'
                  message:
                    type: string
                    example: Account created successfully
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /accounts/{accountNumber}:
    get:
      summary: Get account details
      tags: [Accounts]
      parameters:
        - in: path
          name: accountNumber
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Account details
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    $ref: '#/components/schemas/Account'
        '404':
          description: Account not found
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    
    patch:
      summary: Update account details
      tags: [Accounts]
      parameters:
        - in: path
          name: accountNumber
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: New account name
      responses:
        '200':
          description: Account updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    $ref: '#/components/schemas/Account'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          description: Account not found
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...